// src/app/home/models/tb-pricecf.ts
export class TbPricecf {
  pricecfSerial?: number= null;
  endDate?: Date= null;
	partNo?: string= null;
	priceUnit?: number= null;
	priceUnitx?: number= null;
	pricecfImpSeq?: number= null;
	pricecfImpSerial?: number= null;
	pricecfNo?: string= null;
	remark?: string= null;
	startDate?: Date= null;
	supSerial?: number= null;

	offset?: number= null;
	limit?: number= null;
}